+++ src/main/java/com/woowacourse/pelotonbackend/certification/application/CertificationService.java
@@ -7,7 +7,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.woowacourse.pelotonbackend.certification.domain.Certification;
 import com.woowacourse.pelotonbackend.certification.domain.CertificationRepository;
 import com.woowacourse.pelotonbackend.certification.domain.dto.CertificationCreateRequest;
-import com.woowacourse.pelotonbackend.common.upload.UploadService;
+import com.woowacourse.pelotonbackend.infra.upload.UploadService;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
