+++ src/main/java/com/woowacourse/pelotonbackend/infra/login/KakaoAPIService.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pelotonbackend.member.infra;
+package com.woowacourse.pelotonbackend.infra.login;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
@@ -7,9 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.util.DefaultUriBuilderFactory;
 
-import com.woowacourse.pelotonbackend.member.application.LoginAPIService;
-import com.woowacourse.pelotonbackend.member.infra.dto.KakaoTokenResponse;
-import com.woowacourse.pelotonbackend.member.infra.dto.KakaoUserResponse;
 import com.woowacourse.pelotonbackend.support.dto.JwtTokenResponse;
 import reactor.core.publisher.Mono;
