+++ src/main/java/com/woowacourse/pelotonbackend/infra/upload/UploadService.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pelotonbackend.common.upload;
+package com.woowacourse.pelotonbackend.infra.upload;
 
 import org.springframework.web.multipart.MultipartFile;
