+++ src/main/java/com/woowacourse/pelotonbackend/member/application/MemberService.java
@@ -11,7 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.woowacourse.pelotonbackend.common.exception.MemberIdInvalidException;
 import com.woowacourse.pelotonbackend.common.exception.MemberNotFoundException;
-import com.woowacourse.pelotonbackend.common.upload.UploadService;
+import com.woowacourse.pelotonbackend.infra.upload.UploadService;
 import com.woowacourse.pelotonbackend.member.domain.Member;
 import com.woowacourse.pelotonbackend.member.domain.MemberRepository;
 import com.woowacourse.pelotonbackend.member.presentation.dto.MemberCashUpdateRequest;
