+++ src/test/java/com/woowacourse/pelotonbackend/member/application/LoginServiceTest.java
@@ -11,8 +11,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.woowacourse.pelotonbackend.member.infra.dto.KakaoTokenResponse;
-import com.woowacourse.pelotonbackend.member.infra.dto.KakaoUserResponse;
+import com.woowacourse.pelotonbackend.infra.login.KakaoTokenResponse;
+import com.woowacourse.pelotonbackend.infra.login.KakaoUserResponse;
+import com.woowacourse.pelotonbackend.infra.login.LoginAPIService;
 import com.woowacourse.pelotonbackend.member.presentation.dto.MemberCreateRequest;
 import com.woowacourse.pelotonbackend.support.JwtTokenProvider;
 import com.woowacourse.pelotonbackend.support.RandomGenerator;
